brezee_maa/requirements.txt
numpy
pandas
pandas-ta
duckdb
python-dotenv

pyotp
asyncio
pyppeteer

dhanhq
breeze-connect
fyers-apiv3